requirements.txt
Django==6.0.4
djangorestframework==3.15.2
PyJWT==2.9.0
bcrypt==4.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pytest==8.3.2
pytest-django==4.8.0